# src/routes/auth.ts
import { Router, type Request, type Response } from 'express';
import { SignupSchema, type SignupData, LoginSchema, type LoginData } from '../schemas/auth.js';
import { SuccessResponseSchema, ErrorResponseSchema, type SuccessResponse, type ErrorResponse } from '../schemas/responses.js';
import { authMiddleware } from '../middleware/auth.js';
import { type JWTPayload, type JWTDecoded } from '../schemas/jwt.js';
import { generateJWT } from '../utils/jwt.js';
import { dbService } from '../utils/db.js';

/**
 * Authentication Routes
 * @module routes/auth
 * @description Handles user authentication endpoints: signup, login, and profile retrieval
 */
const authRouter : Router = Router();


/**
 * POST /auth/signup
 * @description Register a new user account
 * @param {SignupData} req.body [body] - User registration data (name, email, password, role)
 * @returns {SuccessResponse} 201 - User created with profile data (no token)
 * @returns {ErrorResponse} 400 - Validation error / email already exists
 * @returns {ErrorResponse} 500 - Failed to create user / Unknown error occurred
 */
authRouter.post('/auth/signup', async (req: Request, res: Response): Promise<void>  => {
  try {
    // Validate request body against SignupSchema
    const validationResult = SignupSchema.safeParse(req.body);
    
    if (!validationResult.success) {
      const errorResponse: ErrorResponse = {
        success: false,
        error: 'Invalid request schema',
      };
      
      ErrorResponseSchema.parse(errorResponse);
      res.status(400).json(errorResponse);
      return;
    }
    
    const signupData: SignupData = validationResult.data;

    // Check if user with the same email already exists
    const existingUser = await dbService.getUserByEmail(signupData.email);
    if (existingUser) {
        const errorResponse: ErrorResponse = {
            success: false,
            error: 'Email already exists',
        };

        ErrorResponseSchema.parse(errorResponse);
        res.status(400).json(errorResponse);
        return;
    }

    // Create new user document
    const newUser = await dbService.createUser(signupData);

    if (newUser === null) {
      const errorResponse: ErrorResponse = {
        success: false,
        error: 'Email already exists',
      };

      ErrorResponseSchema.parse(errorResponse);
      res.status(400).json(errorResponse);
      return;
    }

    const newUserDetails = {
        _id: newUser._id.toString(),
        name: newUser.name,
        email: newUser.email,
        role: newUser.role,
      }

    const successResponse: SuccessResponse = {
      success: true,
      data: newUserDetails,
    };

    SuccessResponseSchema.parse(successResponse);
    res.status(201).json(successResponse);

  } catch (error) {
    // Handle any errors (database, validation, JWT, etc.)
    console.error('❌ Error during signup:', error);
    const errorResponse: ErrorResponse = {
      success: false,
      error: 'Unknown error occurred',
    };

    ErrorResponseSchema.parse(errorResponse);
    res.status(500).json(errorResponse);
  }
});


/**
 * POST /auth/login
 * @description Authenticate user and return JWT token
 * @param {LoginData} req.body [body] - User credentials (email, password)
 * @returns {SuccessResponse} 200 - Authentication successful with JWT token
 * @returns {ErrorResponse} 400 - Invalid email or password / Validation error
 * @returns {ErrorResponse} 500 - Unknown error occurred
 */
authRouter.post('/auth/login', async (req: Request, res: Response): Promise<void> => {
  try {
    // Validate request body against LoginSchema
    const validationResult = LoginSchema.safeParse(req.body);
    if (!validationResult.success) {
      const errorResponse: ErrorResponse = {
        success: false,
        error: 'Invalid request schema',
      };

      ErrorResponseSchema.parse(errorResponse);
      res.status(400).json(errorResponse);
      return;
    }

    const loginData: LoginData = validationResult.data;

    // Find user by email (fetch password for authentication)
    const userData = await dbService.getUserByEmail(loginData.email, true);
    if (userData === null) {
      const errorResponse: ErrorResponse = {
        success: false,
        error: 'Invalid email or password',
      };

      ErrorResponseSchema.parse(errorResponse);
      res.status(400).json(errorResponse);
      return;
    }

    // Compare passwords using UserModel method
    const passwordMatch = await userData.comparePassword(loginData.password);
    if (!passwordMatch) {
      const errorResponse: ErrorResponse = {
        success: false,
        error: 'Invalid email or password',
      };

      ErrorResponseSchema.parse(errorResponse);
      res.status(400).json(errorResponse);
      return;
    }

   // Prepare JWT payload with user credentials
   const jwtPayload: JWTPayload = {
      userId: userData._id.toString(),
      role: userData.role,
    };

    // Generate JWT token
    const jwtToken = generateJWT(jwtPayload);
  
    const successResponse: SuccessResponse = {
      success: true,
      data: {
        token: jwtToken, 
      },
    };

    SuccessResponseSchema.parse(successResponse);
    res.status(200).json(successResponse);

  } catch (error) {
    // Handle other exceptions
    console.error('❌ Error during login:', error);
    const errorResponse: ErrorResponse = {
      success: false,
      error: 'Unknown error occurred',
    };

    ErrorResponseSchema.parse(errorResponse);
    res.status(500).json(errorResponse);
  }
});


/**
 * GET /auth/me
 * @description Retrieve current authenticated user's profile
 * @param {string} req.headers.authorization [header] - JWT token in format <token>
 * @requires Authentication - Must include valid JWT token in Authorization header
 * @returns {SuccessResponse} 200 - Current user's profile data
 * @returns {ErrorResponse} 401 - Unauthorized (invalid or missing token)
 * @returns {ErrorResponse} 404 - User not found
 * @returns {ErrorResponse} 500 - Unknown error occurred
 */
authRouter.get('/auth/me', authMiddleware, async (req: Request, res: Response): Promise<void> => {
  try {
    // authMiddleware ensures req.user exists
    const userData = await dbService.getUserById(req.user!.userId);
    
    if (userData === null) {
      const errorResponse: ErrorResponse = {
        success: false,
        error: 'User not found',
      };
      
      ErrorResponseSchema.parse(errorResponse);
      res.status(404).json(errorResponse);
      return;
    }

    const userDetails = {
        _id: userData._id.toString(),
        name: userData.name,
        email: userData.email,
        role: userData.role,
      }

    const successResponse: SuccessResponse = {
      success: true,
      data: userDetails,
    };

    SuccessResponseSchema.parse(successResponse);
    res.status(200).json(successResponse);
  } catch (error) {
    console.error('❌ Error fetching user profile:', error);
    const errorResponse: ErrorResponse = {
      success: false,
      error: 'Unknown error occurred',
    };

    ErrorResponseSchema.parse(errorResponse);
    res.status(500).json(errorResponse);
  }
});


export default authRouter;