# src/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Attendance API
  version: 1.0.0
  description: API for managing student attendance and classes
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: JWT token - provide token directly without "Bearer" prefix
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
paths:
  /auth/signup:
    post:
      summary: Register a new user account
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name:
                  type: string
                  minLength: 1
                  description: User's full name (required, at least 1 character)
                email:
                  type: string
                  format: email
                  description: User's email address (must be valid email format)
                password:
                  type: string
                  minLength: 6
                  description: User's password (minimum 6 characters required)
                role:
                  type: string
                  enum: [student, teacher]
                  description: User's role (must be either 'student' or 'teacher')
            example:
              name: "Alice Johnson"
              email: "alice.johnson@school.com"
              password: "SecurePass123"
              role: "student"
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  _id: "507f1f77bcf86cd799439011"
                  name: "Alice Johnson"
                  email: "alice.johnson@school.com"
                  role: "student"
        400:
          description: Validation error or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Email already exists"
        500:
          description: Server error creating user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unknown error occurred"
  /auth/login:
    post:
      summary: Authenticate user and return JWT token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  description: User's registered email address
                password:
                  type: string
                  description: User's password
            example:
              email: "alice.johnson@school.com"
              password: "SecurePass123"
      responses:
        200:
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        400:
          description: Validation error or login failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Invalid request schema"
        401:
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Invalid email or password"
        500:
          description: Server error during login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unknown error occurred"
  /auth/me:
    get:
      summary: Get current authenticated user profile
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        200:
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  _id: "507f1f77bcf86cd799439011"
                  name: "Alice Johnson"
                  email: "alice.johnson@school.com"
                  role: "student"
        401:
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unauthorized or invalid token"
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "User not found"
        500:
          description: Server error fetching user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Unknown error occurred"
...